+++ src/Symfony/Component/Form/ChoiceList/Factory/CachingFactoryDecorator.php
@@ -20,7 +20,7 @@ use Symfony\Component\Form\ChoiceList\View\ChoiceListView;
  *
  * @author Bernhard Schussek <bschussek@gmail.com>
  */
-class CachingFactoryDecorator implements ChoiceListFactoryInterface
+class CachingFactoryDecorator implements ExpandedChoiceListFactoryInterface
 {
     /**
      * @var ChoiceListFactoryInterface
@@ -154,19 +154,41 @@ class CachingFactoryDecorator implements ChoiceListFactoryInterface
      */
     public function createView(ChoiceListInterface $list, $preferredChoices = null, $label = null, $index = null, $groupBy = null, $attr = null)
     {
+        return $this->createExpandedView($list, $preferredChoices, $label, $index, $groupBy, $attr);
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function createExpandedView(ChoiceListInterface $list, $preferredChoices = null, $label = null, $index = null, $groupBy = null, $attr = null, $labelAttr = null)
+    {
         // The input is not validated on purpose. This way, the decorated
         // factory may decide which input to accept and which not.
         $hash = self::generateHash(array($list, $preferredChoices, $label, $index, $groupBy, $attr));
 
         if (!isset($this->views[$hash])) {
-            $this->views[$hash] = $this->decoratedFactory->createView(
-                $list,
-                $preferredChoices,
-                $label,
-                $index,
-                $groupBy,
-                $attr
-            );
+            // BC layer
+            if ($this->decoratedFactory instanceof ExpandedChoiceListFactoryInterface) {
+                $choiceView = $this->decoratedFactory->createExpandedView(
+                    $list,
+                    $preferredChoices,
+                    $label,
+                    $index,
+                    $groupBy,
+                    $attr,
+                    $labelAttr
+                );
+            } else {
+                $choiceView = $this->decoratedFactory->createView(
+                    $list,
+                    $preferredChoices,
+                    $label,
+                    $index,
+                    $groupBy,
+                    $attr
+                );
+            }
+            $this->views[$hash] = $choiceView;
         }
 
         return $this->views[$hash];